requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pyjwt
alembic
psycopg2
sqlalchemy
pytest